% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/beer_song.R
\docType{package}
\name{beer_song
Describes Drinking Beer}
\alias{beer_song
Describes Drinking Beer}
\alias{beer_song}
\alias{beer_song
Describes Drinking Beer-package}
\title{This Package Is Just for Practice.
The beer Song function is of no pratical value and is just Used a placeholder.}
\usage{
beer_song(n = 99, vessel = "bottles", liquid = "beer", surface = "wall")
}
\arguments{
\item{n}{integer}

\item{vessel}{character}

\item{liquid}{character}

\item{surface}{character}
}
\description{
This Package Is Just for Practice.
The beer Song function is of no pratical value and is just Used a placeholder.
}
\examples{
beersong(20)
}
